removed a lot of functions that became redundant due to new tracker, tidied up a little, removed coord extend
